+++ waterdata/tests/test_filters.py
@@ -38,27 +38,30 @@ class TestIndefiniteArticleFilter(TestCase):
 class TestDataStartYearFilter(TestCase):
 
     def setUp(self):
-        self.test_series = [
-            {
-                'start_date': datetime.datetime(1889, 2, 28),
-                'end_date': datetime.datetime(2017, 5, 27),
-                'data_types': ['Data Types'],
-                'parameter_code': '00060',
-                'parameter_name': 'Discharge, dm3/s'
-            },
-            {
-                'start_date': datetime.datetime(1890, 12, 1),
-                'end_date': datetime.datetime(2018, 3, 25),
-                'data_types': ['Data Types'],
-                'parameter_code': '00010',
-                'parameter_name': 'Temperature, K'
+        self.test_series = {
+            "Physical": {
+                'parameters': [
+                    {
+                        'start_date': datetime.datetime(1889, 2, 28),
+                        'end_date': datetime.datetime(2017, 5, 27),
+                        'data_types': ['Data Types'],
+                        'parameter_code': '00060',
+                        'parameter_name': 'Discharge, dm3/s'
+                    },
+                    {
+                        'start_date': datetime.datetime(1890, 12, 1),
+                        'end_date': datetime.datetime(2018, 3, 25),
+                        'data_types': ['Data Types'],
+                        'parameter_code': '00010',
+                        'parameter_name': 'Temperature, K'
+                    }
+                ]
             }
-        ]
+        }
 
     def test_series_is_empty(self):
-        result = filters.data_start_year([])
-        expected = 'unknown'
-        self.assertEqual(result, expected)
+        result = filters.data_start_year({})
+        self.assertIsNone(result)
 
     def test_series_has_data(self):
         result = filters.data_start_year(self.test_series)
@@ -69,58 +72,152 @@ class TestDataStartYearFilter(TestCase):
 class TestReadableParmListFilter(TestCase):
 
     def setUp(self):
-        self.test_series = [
+        self.now = datetime.datetime.now()
+        self.recent = datetime.datetime.now() - datetime.timedelta(days=2)
+        self.long_ago = datetime.datetime.now() - datetime.timedelta(days=10)
+        self.physical_series = [
             {
                 'start_date': datetime.datetime(1908, 1, 3),
-                'end_date': datetime.datetime(2016, 3, 23),
-                'data_types': ['Data Types'],
+                'end_date': self.now,
+                'data_types': ['Unit Values'],
                 'parameter_code': '09001',
                 'parameter_name': 'Antimatter, liters'
             },
             {
                 'start_date': datetime.datetime(1908, 1, 3),
-                'end_date': datetime.datetime(2016, 3, 23),
-                'data_types': ['Data Types'],
+                'end_date': self.recent,
+                'data_types': ['Unit Values'],
                 'parameter_code': '09002',
                 'parameter_name': 'Matter, liters'
             },
             {
                 'start_date': datetime.datetime(1908, 1, 19),
-                'end_date': datetime.datetime(2018, 3, 25),
-                'data_types': ['Data Types'],
+                'end_date': self.recent,
+                'data_types': ['Unit Values'],
                 'parameter_code': '00060',
                 'parameter_name': 'Discharge, dm3/s'
             },
             {
                 'start_date': datetime.datetime(1908, 1, 19),
-                'end_date': datetime.datetime(2018, 3, 25),
-                'data_types': ['Data Types'],
+                'end_date': self.recent,
+                'data_types': ['Unit Values'],
                 'parameter_code': '00010',
                 'parameter_name': 'Temperature, K'
+            },
+            {
+                'start_date': datetime.datetime(1908, 1, 19),
+                'end_date': self.recent,
+                'data_types': ['Daily Values'],
+                'parameter_code': '22140',
+                'parameter_name': 'Turbidity'
+            },
+            {
+                'start_date': datetime.datetime(1908, 1, 19),
+                'end_date': self.recent,
+                'data_types': ['Daily Values'],
+                'parameter_code': '22140',
+                'parameter_name': 'Depth'
+            }
+        ]
+        self.inorganic_series = [
+            {
+                'start_date': datetime.datetime(1908, 1, 3),
+                'end_date': datetime.datetime.now(),
+                'data_types': ['Unit Values', 'Daily Values'],
+                'parameter_code': '07242',
+                'parameter_name': 'Lead, g/mL'
+            },
+            {
+                'start_date': datetime.datetime(1999, 1, 26),
+                'end_date': self.recent,
+                'data_types': ['Unit Values', 'Daily Values'],
+                'parameter_code': '02220',
+                'parameter_name': 'Neptunium, ug/mL'
+            }
+        ]
+        self.old_inorganic_series = [
+            {
+                'start_date': datetime.datetime(1908, 1, 3),
+                'end_date': self.long_ago,
+                'data_types': ['Unit Values', 'Daily Values'],
+                'parameter_code': '07242',
+                'parameter_name': 'Lead, g/mL'
+            },
+            {
+                'start_date': datetime.datetime(1999, 1, 26),
+                'end_date': self.long_ago,
+                'data_types': ['Unit Values', 'Daily Values'],
+                'parameter_code': '02220',
+                'parameter_name': 'Neptunium, ug/mL'
             }
         ]
 
     def test_series_is_empty(self):
-        result = filters.readable_param_list([])
-        expected = 'unknown'
-        self.assertEqual(result, expected)
+        result = filters.readable_param_list({})
+        self.assertIsNone(result)
 
     def test_single_measured_parameter(self):
-        result = filters.readable_param_list(self.test_series[:1])
+        result = filters.readable_param_list({'Physical': {
+            'parameters': self.physical_series[:1]
+        }})
         expected = 'ANTIMATTER'
         self.assertEqual(result, expected)
 
     def test_two_measured_parameters(self):
-        result = filters.readable_param_list(self.test_series[:2])
+        result = filters.readable_param_list({'Physical': {
+            'parameters': self.physical_series[:2]
+        }})
         expected = 'ANTIMATTER and MATTER'
         self.assertEqual(result, expected)
 
     def test_three_measured_parameters(self):
-        result = filters.readable_param_list(self.test_series[:3])
+        result = filters.readable_param_list({'Physical': {
+            'parameters': self.physical_series[:3],
+            'data_types': 'Unit Values, Water Quality',
+            'end_date': datetime.datetime.now()
+        }})
         expected = 'DISCHARGE, ANTIMATTER, and MATTER'
         self.assertEqual(result, expected)
 
     def test_four_measured_parameters(self):
-        result = filters.readable_param_list(self.test_series)
+        result = filters.readable_param_list({'Physical': {
+            'parameters': self.physical_series,
+            'data_types': 'Unit Values, Water Quality',
+            'end_date': datetime.datetime.now()
+        }})
         expected = 'DISCHARGE, TEMPERATURE, ANTIMATTER, and MORE'
         self.assertEqual(result, expected)
+
+    def test_no_current_data(self):
+        result = filters.readable_param_list({'Inorganic': {
+            'parameters': self.old_inorganic_series,
+        }})
+        self.assertIsNone(result)
+
+    def test_no_unit_values(self):
+        result = filters.readable_param_list({'Physical': {
+            'parameters': self.physical_series[-2:]
+        }})
+        self.assertIsNone(result)
+
+    def test_mixed_data_types(self):
+        result = filters.readable_param_list({
+            'Physical': {
+                'parameters': self.physical_series[-2:]
+            },
+            'Inorganic': {
+                'parameters': self.inorganic_series,
+            }
+        })
+        expected = 'LEAD and NEPTUNIUM'
+        self.assertEqual(result, expected)
+
+
+class TestDateToStringFilter(TestCase):
+
+    def setUp(self):
+        self.test_dt = datetime.datetime(2018, 3, 25)
+
+    def test_year_month_day(self):
+        result = filters.date_to_string(self.test_dt)
+        self.assertEqual(result, '2018-03-25')