+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Timeseries SVG has the correct title and desc tag contents.
 - Series with no points are not considered when determining the default parameter.
+- Remove source maps from CSS build
+- Include only real-time parameters with recent data in a site's description
+- Tooltip points are not created for points containing infinite y-axis value
 
 
 ## [0.5.0] - 2018-03-30